load average now estimated in simulation
